library: implement waiting for activation/deactivation with NetworkManager
Implement what nmcli does when activating/deactivating
a connection: waits until the device is fully activated
or deactivate.

The 'wait' option is basically like nmcli's --wait option.
If omitted, wait a default duration of time. If the action
does not complete within the specified timeout, it fails.
Note that a failure may be suppressed by setting 'ignore_errors',
either globaly or per-state.

Set 'wait=0' not to wait. This was the previous behavior before
this commit.