library: autodetect the "slave_type" property based on "master"
As it is, a slave profile MUST reference an existing "master"
profile. Hence, the "slave_type" is implicitly known and not
required.

However, don't drop the "slave_type" option entirely. It allows a user
to explicitly set it, and network_connections.py will reject mismatching
configurations.

Also, check that the 'master' profile is actually of a valid master
type for the slave.