role: improve making "network_provider" configurable via host vars
The role supports two providers: "nm" and "initscripts".

The provider is autodetected by loading one of the vars/*.yml files
(where the default is loaded as internal "network_provider_default" variable).
The user can still overwrite the provider, by explicitly setting the
"network_provider" variable.

Depending on the provider there is the list of packages that shall be
installed and the service to start. Selecting this was broken before.

This is now fixed and works like following:

The variables "network_service_name" and "network_packages" can be
specified by the user as host variables. But usually one wouldn't
want to do that. In that case, we look into the internal
"_network_provider_setup" dictionary, which uses the variables
"network_service_name_nm", "network_service_name_initscripts",
"network_packages_nm", "network_packages_initscripts".

These variables are initialized in "defaults/main.yml" as well,
but they could be overwritten via "vars/*.yml" files.

https://bugzilla.redhat.com/show_bug.cgi?id=1485074