library: add provider specifiy pre-validation step of input
Depending on the provider, we have additional restrictions on the
input arguments. Validate them early.

Without this, we will only fail later when we want to get the
'interface_name' of a parent or a master profile. Fail early.

The major reason for this, is to expose this validation so that
it can be used by unit tests, to check whether proceeding will
lead to a known failure.