Unset the executable bit on the unit test.
This script is not being executed directly, but by using python, python2 or
python3 interpreter and passing the script name as argument. This is because
the script is to be tested under multiple versions of Python. (It is also only
referenced in a copy task and the remote copy gets executed, which makes the
executable bit on the original unneeded as well.)

(Some RPM checks
https://src.fedoraproject.org/rpms/redhat-rpm-config/blob/f7e8f73eadbd09f6eee0e836bf3cb80f14f0c053/f/brp-mangle-shebangs
do not like an executable script to have a #!/usr/bin/python shebang. This
script has it to indicate that it is compatible with both versions of Python,
although a shebang line is not very meaningful here for the same reason as the
executable bit.)