module: rework polling in NMUtil.wait_till_connection_is_gone()
time.sleep() does not sleep an exact amount of time, hence, adding
up the sleep-times will not give the exact total wait-time.
An alternative would be to take a CLOCK_BOOTTIME timestamp (or similar)
and determine the elapsed time based on that.

Instead, do something different, and use a GLib timeout for polling.