Allow more BONDING_OPTS
At the moment the role supports configuring the BONDING_OPTS parameters `mode` and `miimon`, however a great variety of parameters might also be supported.

This PR adds support for `use_carrier`, `updelay`, `downdelay`, `primary_reselect` and `primary`. Appreciate that's still only a small subset so I understand if you'd rather reject and wait for someone to do a fuller job, but these are the ones I needed for my project so thought I'd share.

Probably the biggest gap is that `primary` refers to the name of another network_connection. I've left that direct (i.e., you specify the `interface_name`) but I guess you might want to match the behaviour you have for `master` which takes the Ansible profile name and converts.