sync with latest template
Just use the `python` command in the virtual environment, instead
of using `envpython`.  This solves some problems where the
attempt to resolve the `envpython` to the `syspython` by following
symlinks causes python not to be able find some packages installed
in the venv.

Add support for python 3.8-dev in travis.

Do not install ansible by default in all tox venvs for pytest
tests.  If the user needs to install ansible for pytests, the
user will add `ansible` to the new `pytest_extra_requirements.txt`
For python 2.6, which requires ansible 2.6, there is a special
`pytest26_extra_requirements.txt`, and an example
`ansible26_requirements.txt` which the user will use to specify
the extra dependencies for installing ansible 2.6 on python 2.6.

There are two extra dependencies added to the list of dependencies
installed on the travis machines when using python 2.6 - libffi-dev
and libssl-dev - these are required in order to install ansible 2.6
and its extra dependencies.

For python 2.6, an older version of tox must be used, which does not
have support for `commands_pre`, so allow `tests/setup_module_utils.sh`
to be run directly from `.travis/runpytest.sh` on those platforms.
the user will select to install ansible

Use the file `.yamllint_defaults.yml` for the base `yamllint`
configuration.  User customizations go into `.yamllint.yml`, which
extends `.yamllint_defaults.yml`.  This allows the use of `yamllint`
with IDEs that do not allow specifying command line arguments
for the `yamllint` command.

Only run coveralls in travis in the py35 environment.