tests_unit.yml - support the collection converted format.
The unit test depends on the relative paths to library and module_utils,
which are no longer available from the tests dir. (The tests directory
in the collections CI is independent from the collections path.) To
workaround it, the library and module_utils are copied to the tests
directory and let the unit test additionally refer them.

Signed-off-by: Noriko Hosoi <nhosoi@redhat.com>