tests_integration_pytest.yml - support the collection converted format.
To run the pytest on the target hosts via CI tests, integration_pytest_
python3.yml copies entire network role to the target. To get the top
directory, it uses a git command line. But the method cannot be used
in the collection converted format since the converted roles and tests
are not located in the git-cloned directory.

In the collection converted format, the network tests directory has the
copy of modules and module_utils. Taking advantage of the self-contained
tests, we can just copy the files and directories in the tests and run
test_integration_pytest successfully in the target host.

Signed-off-by: Noriko Hosoi <nhosoi@redhat.com>