Support specifying mac address without quotation mark
- In yaml, all the string value consists of a series of one-digit or
two-digit numbers delimited by colons and all numbers but the first are
between 0 and 59, it would be interpreted as a sexagesimal number, and
would be converted automatically to the equivalent number of seconds.

- The committed patch will guarantee to remediate the yaml's peculiarity
and convert back into the mac address specified in the original yaml
file. And in case that user specifies arbitrary integer(e.g. `mac:1234`),
we introduce `force_len`(e.g. `MAC_ADDR_OCTETS = [6, 20]`) to raise
error for invalid integer value.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>