tag 802-1x, bond tests as expfail; bridge test cleanup
The 802-1x test will fail on platforms where `hostapd` is not available,
so tag that test to make it skippable.

The initscripts bridge test does not clean up properly, leaving the
device around which causes the nm test to fail.  Explicitly remove
the device for cleanup.

The bond tests do not work when being run sequentially on the same
machine as in the smoke test scenario, so tag those tests so that
they can be skipped.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>