Reload the connection before managing the device
For the test playbook `tests_ethernet.yml`, when initscripts provider
runs first, the connection profile contains `NM_CONTROLLed=no`, which
indicates that the associated veth device is not managed by NM, and the
veth device and connection profile get cleaned up at the end. When NM
provider runs, the veth device stay unmanaged even if we explicitly set
it to managed because the NM is not detecting the change of connection
profile without reloading the connection.

To fix it, reload the connection before managing the device when NM
provider runs.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>