Bring up ports when testing bond in Fedora
When running the bond integration tests with initscripts provider over
Fedora target hosts, the carrier flag of the controller device can not be
set if the port devices are not brought up. As a result, the controller
device can not be brought up.

To fix it, bring up ports before activating the connection profiles on
the controller.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>