Instead of set_fact, use vars to set "interface"
This is part of the effort to allow CI tests run in the serialized
manner on one VM for shortening the duration of the CI tests. If
set_fact is used, a value of "interface" in a test is visible to
the next test, which could make the next test fail.

Plus this commit adds a network restart code in bond_deprecated.
After running bond_deprecated test case, the network is down.
Unless it's restarted the following test fails in the necessary
package download.