infiniband: Fix the bug of wrongly checking whether the device exists
When the `p_key` is set explicitly, the virtual infiniband device
should be created automatically. And activating the connection on the
virtual infiniband device should not require the users to create
virtual infiniband device manually upfront. The previous implementation
is wrong which validating the presence of virtual infiniband device
upfront, as a result, the users have to create the virtual infiniband
device manually in the target system in order to activate the
connection on the virtual infiniband device through the role. This is
an actual bug, the exception raised when the users do not
pre-configured such a virtual infiniband device. To address the
problem, only checking the existence of the device when the `p_key` is
-1, which means that the `p_key` is not set explicitly.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>