Add support for __network_test_flags to workaround CI problems
On some CI systems the tests that create routes can create them in such
a way that they override the "default" default routes, interrupting
network connectivity and causing Ansible to hang.  The parameter
`__network_test_flags` gives network role developers a way to ensure
that the newly created route is given a lower priority and does not
override the "default" default route.  A `dict` was chosen as the type
of this variable so that it could be easily extended in the future for
future use cases.