ansible-lint: Remediate the test failures
Previously we just exlude the ansible-lint check on the `./test`
directory, but it will veil all the test failures. To properly address
the ansible-lint failures (unamed-tasks, command-instead-of-module,
yaml[truthy] etc), we can adjust the tests or append `# noqa` to the
tasks or update the warn_list and skip_list to suppress the errors.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>