ansible-lint: Define variable 'network_provider' to fix jinja[invalid] failure
ansible-lint 6.x syntax check raises `jinja[invalid]` error on the
mandatory variable 'network provider' because it is not defined. In
order to fix the problem, define the 'network_provider' variable in
the file .ansible-lint. The value of the variable is less important as
long as it is reasonable which only meets the purpose of passing the
ansible-lint 6.x syntax check.

Signed-off-by: Wen Liang <liangwen12year@gmail.com>